requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lyricsgenius==3.0.1
MarkupSafe==2.0.1
nltk==3.6.3
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.9.24
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.2.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.1
